+++ src/opinf/lstsq/_total.py
